add RuntimeError to results, don't raise exception
Errors may be included at some point in the response although other data is available.
SQL ERROR can already happen when a player does not exist, so it doesn't make a lot of sense to raise an exception if only one of many results is not available.

It is now the user's job to check for RuntimeErrors in the result lists.